refactor([issue-3174]): extract sw.js test sandbox to swSandbox.js
Follows the client/src/test/ convention (mockEventSource.js, fakeAudioContext.js)
of factoring reusable fake-environment helpers into their own module so a
future sw.js test (install/activate precache, cacheFirst, etc.) doesn't have
to duplicate the vm sandbox. Also drops two return fields nothing consumed.
